added a BaseEntity for models

+++ pet-clinic-data/src/main/java/guru/framework/petclinic/services/OwnerService.java
@@ -5,7 +5,7 @@ import java.util.Set;
 import guru.framework.petclinic.model.Owner;
 
 public interface OwnerService {	
-	Owner findById(Long id);
+	Owner findById(Integer id);
 	Set<Owner> findAll();
 	Set<Owner> findByLastName(String lastName);
 	Owner save(Owner owner);
