+++ pet-clinic-data/src/main/java/guru/framework/petclinic/services/PetService.java
@@ -5,7 +5,7 @@ import java.util.Set;
 import guru.framework.petclinic.model.Pet;
 
 public interface PetService {
-	Pet findById(Long id);
+	Pet findById(Integer id);
 	Set<Pet> findAll();
 	Pet save(Pet pet);
 }
